upd
git-svn-id: file:///Users/demichel/SVNREPOS/CODE@4983 fea3b138-1bf5-0310-9da5-cbb00e418fd0

+++ MDsimul++/hard_ellipses/expand.cpp
@@ -3,7 +3,7 @@ using ntype = long double;
 int main(int argc, char **argv)
 { 
   double fx=1.0, fy=1.0;
-  mcsim<double> heMC;
+  mcsim<long double> heMC;
 
   if (argc==1)
     {
